Moving Merkle tree to a separate package to remove zei_api dependency on storage
