+++ accumulators/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "accumulators"
+version = "0.1.0"
+authors = ["Findora <engineering@findora.org>"]
+edition = "2018"
+description = "Zei accumulators commons"
+
+[lib]
+name = "accumulators"
+crate-type = ["rlib"]
+
+[dependencies]
+algebra = { path = "../algebra" }
+crypto = { path = "../crypto" }
+itertools = "0.8.0"
+storage = { git = "https://github.com/FindoraNetwork/storage.git", branch = "develop" }
+ruc = "0.6.2"
+utils = { path = "../utils" }
+parking_lot = "0.11.2"
