+++ zei_api/Cargo.toml
@@ -14,6 +14,7 @@ name = "zei"
 crate-type = ["rlib"]
 
 [dependencies]
+accumulators = { path = "../accumulators" }
 algebra = { path = "../algebra" }
 boolinator = "2.4.0"
 bulletproofs = { package = "bulletproofs", git = "https://github.com/FindoraNetwork/bp", branch = "batch_verification", features = ["yoloproofs"] }
@@ -37,7 +38,6 @@ poly-iops = { path = "../poly-iops" }
 structopt = "0.3.17"
 bincode = "1.3.1"
 ruc = "0.6.2"
-storage = { git = "https://github.com/FindoraNetwork/storage.git", branch = "develop" }
 parking_lot = "0.11.1"
 
 [dev-dependencies]
@@ -48,7 +48,7 @@ lazy_static = "1.4.0"
 rand = { version = "0.7", default-features = false }
 serde_json = "1.0"
 typenum = "1.11.2"
-bench-utils = { path = "../bench-utils" }
+storage = { git = "https://github.com/FindoraNetwork/storage.git", branch = "develop" }
 
 [features]
 default = ["std", "u64_backend"]
